// src/app/shared/infrastructure/api/generated/model/authResponseResource.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Backend-WebApplication
 * ChapaTuRuta REST API Ã¢ÂÂ Identity & Access Management, Driver Management, and Monetization
 * OpenAPI spec version: 1.0.0
 */
import type { ProfileResource } from './profileResource';
import type { UserResource } from './userResource';

/**
 * Response body for authentication and registration
 */
export interface AuthResponseResource {
  user?: UserResource;
  profile?: ProfileResource;
  /** JWT access token (15-minute TTL) */
  accessToken?: string;
  /** Opaque refresh token (7-day TTL) */
  refreshToken?: string;
}